include slider for deleting sessions

+++ Extension/ui/views/DeleteSessionConfirmationView.swift
@@ -1,31 +1,45 @@
 import SwiftUI
+import CompactSlider
 
-func deleteSessionConfirmationView(geometry: GeometryProxy, store: Store) -> some View {
-    VStack {
-        Text("Delete session from " + store.state.selectedSessionId + "?")
-            .font(.system(size: 12))
-            .frame(maxHeight: .infinity, alignment: .center)
-        
-        HStack {
-            Button(action: {
-                deleteSession(store: store, sessionId: store.state.selectedSessionId)
-            }) {
-                Text("Delete")
-            }
-            .font(.system(size: 12))
-            .fontWeight(.bold)
-            .buttonStyle(.bordered)
-            .tint(colorize("red"))
+struct DeleteSessionConfirmationView: View {
+    var geometry: GeometryProxy
+    var store: Store
 
-            Button(action: {
-                onLogSelect(store: store)
-            }) {
-                Text("Cancel")
+    @State var value: Double = CONFIRMATION_DEFAULT_VALUE
+
+    init(geometry: GeometryProxy, store: Store) {
+        self.geometry = geometry
+        self.store = store
+    }
+
+    var body: some View {
+        VStack {
+            CompactSlider(value: $value, handleVisibility: .hidden) {
+                Text("Delete session?")
+                    .font(.system(size: 12))
             }
-            .font(.system(size: 12))
-            .fontWeight(.bold)
-            .buttonStyle(.bordered)
-            .tint(colorize("teal"))
-        }.frame(alignment: .center)
-    }.frame(maxHeight: .infinity)
+            .onChange(of: value, perform: {_ in
+                if value > CONFIRMATION_ENOUGH_VALUE {
+                    deleteSession(store: store, sessionId: store.state.selectedSessionId)
+                }
+            })
+            .compactSliderStyle(CustomCompactSliderStyle())
+            .frame(maxHeight: .infinity, alignment: .bottom)
+            .edgesIgnoringSafeArea(.all)
+
+            HStack {
+                Button(action: {
+                    onLogSelect(store: store)
+                }) {
+                    Text("Cancel")
+                }
+                .font(.system(size: 12))
+                .fontWeight(.bold)
+                .buttonStyle(.bordered)
+                .tint(colorize("teal"))
+            }
+            .frame(maxHeight: .infinity, alignment: .bottom)
+            .edgesIgnoringSafeArea(.all)
+        }
+    }
 }
